+++ Cargo.toml
@@ -9,16 +9,23 @@ env_logger = "0.5.10"
 itertools = "0.7"
 log = "0.4.1"
 merkle = { git = "https://github.com/vkomenda/merkle.rs", branch = "public-proof" }
-protobuf = "1.6.0"
+protobuf = { version = "1.6.0", optional = true }
 reed-solomon-erasure = "3.0"
 ring = "^0.12"
 serde = "1.0.54"
 
+[features]
+serialization-protobuf = [ "protobuf", "protoc-rust" ]
+
 [build-dependencies]
-protoc-rust = "1.6.0"
+protoc-rust = { version = "1.6.0", optional = true }
 
 [dev-dependencies]
 crossbeam = "0.3.2"
 crossbeam-channel = "0.1"
 docopt = "0.8"
 rand = "0.3"
+
+[[example]]
+name = "consensus-node"
+required-features = [ "serialization-protobuf" ]
