+++ src/lib.rs
@@ -10,6 +10,7 @@ extern crate bincode;
 extern crate log;
 extern crate itertools;
 extern crate merkle;
+#[cfg(feature = "serialization-protobuf")]
 extern crate protobuf;
 extern crate reed_solomon_erasure;
 extern crate ring;
@@ -20,5 +21,8 @@ pub mod broadcast;
 pub mod common_subset;
 pub mod honey_badger;
 pub mod messaging;
+#[cfg(feature = "serialization-protobuf")]
 pub mod proto;
+#[cfg(feature = "serialization-protobuf")]
 pub mod proto_io;
+mod util;
